closes #320 - init commit, makes Alert mutations an admin-only request

+++ helpers/controllers/alertController.ts
@@ -1,6 +1,7 @@
 import db from '../dbload'
 import { LoggedRequest } from '../../@types/helpers'
 
+import _ from 'lodash'
 const { Alert } = db
 
 type AlertData = {
@@ -16,7 +17,11 @@ export const addAlert = async (
   ctx: { req: LoggedRequest }
 ) => {
   const { req } = ctx
+  const isAdmin = _.get(req, 'user.isAdmin', false)
   try {
+    if (isAdmin === 'false') {
+      throw new Error('User is not an admin')
+    }
     const { text, type, url, urlCaption } = arg
     if (!text || !type) {
       throw new Error('Missing alert parameters')
@@ -37,7 +42,11 @@ export const removeAlert = async (
   ctx: { req: LoggedRequest }
 ) => {
   const { req } = ctx
+  const isAdmin = _.get(req, 'user.isAdmin', false)
   try {
+    if (isAdmin === 'false') {
+      throw new Error('User is not an admin')
+    }
     const { id } = arg
     await Alert.destroy({ where: { id } })
     return {
